requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-tavily>=0.0.3
langchainhub>=0.1.20
python-dotenv>=1.0.0